+++ src/torch_cubic_spline_grids/_constants.py
