+++ tests/test_grid_optimisation.py
@@ -1,7 +1,7 @@
 import einops
 import torch
 
-from torch_cubic_b_spline_grid import (
+from torch_cubic_spline_grids import (
     CubicBSplineGrid1d,
     CubicBSplineGrid2d,
     CubicBSplineGrid3d,
